+++ cmd/gpud/list-plugins/command.go
@@ -1,4 +1,4 @@
-package command
+package listplugins
 
 import (
 	"context"
@@ -7,32 +7,32 @@ import (
 
 	"github.com/urfave/cli"
 
-	v1 "github.com/leptonai/gpud/client/v1"
+	clientv1 "github.com/leptonai/gpud/client/v1"
 	"github.com/leptonai/gpud/pkg/config"
 	"github.com/leptonai/gpud/pkg/log"
 )
 
-// cmdListPlugins implements the list-plugins command
-func cmdListPlugins(c *cli.Context) error {
-	// Set up logging
+// Command implements the list-plugins command
+func Command(cliContext *cli.Context) error {
+	logLevel := cliContext.String("log-level")
 	zapLvl, err := log.ParseLogLevel(logLevel)
 	if err != nil {
 		return err
 	}
-	log.Logger = log.CreateLogger(zapLvl, logFile)
+	log.Logger = log.CreateLogger(zapLvl, "")
 
 	// Create a context with timeout
 	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Minute)
 	defer cancel()
 
 	// Get the server address from the flag, default to http://localhost:<Default GPUd port>
-	serverAddr := c.String("server")
+	serverAddr := cliContext.String("server")
 	if serverAddr == "" {
 		serverAddr = fmt.Sprintf("https://localhost:%d", config.DefaultGPUdPort)
 	}
 
 	// Get custom plugins
-	plugins, err := v1.GetCustomPlugins(ctx, serverAddr)
+	plugins, err := clientv1.GetCustomPlugins(ctx, serverAddr)
 	if err != nil {
 		return fmt.Errorf("failed to get custom plugins: %w", err)
 	}
